create filter package
