+++ stac_fastapi/sfeos_helpers/stac_fastapi/sfeos_helpers/filter/cql2.py
@@ -1,0 +1,35 @@
+"""CQL2 pattern conversion helpers for Elasticsearch/OpenSearch."""
+
+import re
+
+from stac_fastapi.core.extensions.filter import (
+    cql2_like_patterns,
+    valid_like_substitutions,
+)
+
+
+def _replace_like_patterns(match: re.Match) -> str:
+    pattern = match.group()
+    try:
+        return valid_like_substitutions[pattern]
+    except KeyError:
+        raise ValueError(f"'{pattern}' is not a valid escape sequence")
+
+
+def cql2_like_to_es(string: str) -> str:
+    """
+    Convert CQL2 "LIKE" characters to Elasticsearch "wildcard" characters.
+
+    Args:
+        string (str): The string containing CQL2 wildcard characters.
+
+    Returns:
+        str: The converted string with Elasticsearch compatible wildcards.
+
+    Raises:
+        ValueError: If an invalid escape sequence is encountered.
+    """
+    return cql2_like_patterns.sub(
+        repl=_replace_like_patterns,
+        string=string,
+    )
